Added AbstractMultiJobHandler as place-holder super class of two instantiation handlers.
